requirements.txt
Flask
spacy
paho-mqtt